Update deps
- pug@^2.0.0-beta6
- fs-extra removed
